# apps/api/openapi.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Soonr API",
    "description": "Fastify backend for Soonr application routes, interactive docs, and generated client contracts.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/notifications/unread-count": {
      "get": {
        "summary": "Get unread notification count",
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "unreadCount"
                  ],
                  "properties": {
                    "unreadCount": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/notifications": {
      "get": {
        "summary": "List notifications",
        "tags": [
          "notifications"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "titleId",
                          "eventType",
                          "destinationKind",
                          "destinationTitleId",
                          "titleName",
                          "titleArtworkUrl",
                          "message",
                          "subtitle",
                          "payload",
                          "createdAt",
                          "readAt"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "titleId": {
                            "type": "string"
                          },
                          "eventType": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "release_date_changed"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "release_approaching"
                                ]
                              }
                            ]
                          },
                          "destinationKind": {
                            "type": "string",
                            "enum": [
                              "title"
                            ]
                          },
                          "destinationTitleId": {
                            "type": "string"
                          },
                          "titleName": {
                            "type": "string"
                          },
                          "titleArtworkUrl": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "message": {
                            "type": "string"
                          },
                          "subtitle": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "payload": {
                            "anyOf": [
                              {
                                "type": "object",
                                "required": [
                                  "previousReleaseDate",
                                  "nextReleaseDate"
                                ],
                                "properties": {
                                  "previousReleaseDate": {
                                    "anyOf": [
                                      {
                                        "type": "string"
                                      },
                                      {
                                        "type": "null"
                                      }
                                    ]
                                  },
                                  "nextReleaseDate": {
                                    "anyOf": [
                                      {
                                        "type": "string"
                                      },
                                      {
                                        "type": "null"
                                      }
                                    ]
                                  }
                                }
                              },
                              {
                                "type": "object",
                                "required": [
                                  "targetReleaseDate",
                                  "timingPreset"
                                ],
                                "properties": {
                                  "targetReleaseDate": {
                                    "anyOf": [
                                      {
                                        "type": "string"
                                      },
                                      {
                                        "type": "null"
                                      }
                                    ]
                                  },
                                  "timingPreset": {
                                    "anyOf": [
                                      {
                                        "type": "string",
                                        "enum": [
                                          "on_day"
                                        ]
                                      },
                                      {
                                        "type": "string",
                                        "enum": [
                                          "hours_24_before"
                                        ]
                                      },
                                      {
                                        "type": "string",
                                        "enum": [
                                          "days_7_before"
                                        ]
                                      },
                                      {
                                        "type": "string",
                                        "enum": [
                                          "days_30_before"
                                        ]
                                      }
                                    ]
                                  }
                                }
                              }
                            ]
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "readAt": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        }
                      }
                    },
                    "nextCursor": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/notifications/read": {
      "post": {
        "summary": "Mark a notification as read",
        "tags": [
          "notifications"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "notificationId"
                ],
                "properties": {
                  "notificationId": {
                    "minLength": 1,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "notification"
                  ],
                  "properties": {
                    "notification": {
                      "type": "object",
                      "required": [
                        "id",
                        "titleId",
                        "eventType",
                        "destinationKind",
                        "destinationTitleId",
                        "titleName",
                        "titleArtworkUrl",
                        "message",
                        "subtitle",
                        "payload",
                        "createdAt",
                        "readAt"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "titleId": {
                          "type": "string"
                        },
                        "eventType": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "release_date_changed"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "release_approaching"
                              ]
                            }
                          ]
                        },
                        "destinationKind": {
                          "type": "string",
                          "enum": [
                            "title"
                          ]
                        },
                        "destinationTitleId": {
                          "type": "string"
                        },
                        "titleName": {
                          "type": "string"
                        },
                        "titleArtworkUrl": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "subtitle": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "payload": {
                          "anyOf": [
                            {
                              "type": "object",
                              "required": [
                                "previousReleaseDate",
                                "nextReleaseDate"
                              ],
                              "properties": {
                                "previousReleaseDate": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ]
                                },
                                "nextReleaseDate": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ]
                                }
                              }
                            },
                            {
                              "type": "object",
                              "required": [
                                "targetReleaseDate",
                                "timingPreset"
                              ],
                              "properties": {
                                "targetReleaseDate": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ]
                                },
                                "timingPreset": {
                                  "anyOf": [
                                    {
                                      "type": "string",
                                      "enum": [
                                        "on_day"
                                      ]
                                    },
                                    {
                                      "type": "string",
                                      "enum": [
                                        "hours_24_before"
                                      ]
                                    },
                                    {
                                      "type": "string",
                                      "enum": [
                                        "days_7_before"
                                      ]
                                    },
                                    {
                                      "type": "string",
                                      "enum": [
                                        "days_30_before"
                                      ]
                                    }
                                  ]
                                }
                              }
                            }
                          ]
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "readAt": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/notifications/read-all": {
      "post": {
        "summary": "Mark all notifications as read",
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "markedCount"
                  ],
                  "properties": {
                    "markedCount": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local API"
    }
  ]
}